Data synching between devices
Wasn't 100% sure what format the data would be in (what was sent to the
watch) but I set it up to send a data map so you can add whatever you
want to it. Once I have your code I can set it up better to collect and
send the right things as well. Let me know if I have missed something
and I can update it ASAP.